+++ lib/presentation/Home/screen_home.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_movie_app/core/colors/colors.dart';
+import 'package:flutter_movie_app/presentation/Home/widget/custom_icon_widget.dart';
 import 'package:flutter_movie_app/presentation/Home/widget/number_tile_card.dart';
 
 import '../../core/const.dart';
@@ -13,44 +14,7 @@ class ScreenHome extends StatelessWidget {
     return Scaffold(
       body: ListView(
         children: [
-          Stack(
-            children: [
-              Container(
-                width: double.infinity,
-                height: 600,
-                decoration: const BoxDecoration(
-                  image: DecorationImage(
-                    fit: BoxFit.cover,
-                    image: NetworkImage(kMainImg),
-                  ),
-                ),
-              ),
-              Positioned(
-                bottom: 0,
-                left: 0,
-                right: 0,
-                child: Row(
-                  mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                  children: [
-                    Column(
-                      children: const [
-                        Icon(
-                          Icons.add,
-                          color: kWhite,
-                          size: 28,
-                        ),
-                        Text(
-                          'My List',
-                          style: TextStyle(fontSize: 20, color: kWhite),
-                        ),
-                      ],
-                    ),
-                    _playButton(),
-                  ],
-                ),
-              )
-            ],
-          ),
+          
           const MainTitleCard(
             title: 'Released in Past year',
           ),
@@ -76,26 +40,5 @@ class ScreenHome extends StatelessWidget {
     );
   }
 
-  TextButton _playButton() {
-    return TextButton.icon(
-      style: ButtonStyle(
-        backgroundColor: MaterialStateProperty.all(
-          kWhite,
-        ),
-      ),
-      onPressed: () {},
-      icon: const Icon(
-        Icons.play_arrow,
-        size: 30,
-        color: kBlack,
-      ),
-      label: const Padding(
-        padding: EdgeInsets.symmetric(horizontal: 10),
-        child: Text(
-          "Play",
-          style: TextStyle(fontSize: 23, color: kBlack),
-        ),
-      ),
-    );
-  }
+  
 }
